Dump transaction version as an unsigned integer in RPC/TxToUniv
Summary:
> Dump transaction version as an unsigned integer in RPC/TxToUniv
>
> Consensus-wise we already treat it as an unsigned integer (the
> only rules around it are in CSV/locktime handling), but changing
> the underlying data type means touching consensus code for a
> simple cleanup change, which isn't really worth it.
>
> See-also, https://github.com/rust-bitcoin/rust-bitcoin/pull/299

> Additionally treat Tx.nVersion as unsigned in joinpsbts
>
> This gets its own release note callout, though doesn't appear to
> violate the BIP as the BIP appears to be underspecified. We
> probably want to update BIP 174 to mention how version numbers are
> combined.

This is a backport of [[https://github.com/bitcoin/bitcoin/pull/16525 | core#16525]]

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D10013

+++ src/core_write.cpp
@@ -216,7 +216,10 @@ void TxToUniv(const CTransaction &tx, const uint256 &hashBlock, UniValue &entry,
               bool include_hex, int serialize_flags) {
     entry.pushKV("txid", tx.GetId().GetHex());
     entry.pushKV("hash", tx.GetHash().GetHex());
-    entry.pushKV("version", tx.nVersion);
+    // Transaction version is actually unsigned in consensus checks, just
+    // signed in memory, so cast to unsigned before giving it to the user.
+    entry.pushKV("version",
+                 static_cast<int64_t>(static_cast<uint32_t>(tx.nVersion)));
     entry.pushKV("size", (int)::GetSerializeSize(tx, PROTOCOL_VERSION));
     entry.pushKV("locktime", (int64_t)tx.nLockTime);
